:new: Build sources with rollup
